Agregar clase Contacto abstracta
